generalize lighting

+++ syrillian_utils/src/lib.rs
@@ -3,14 +3,9 @@ use std::fmt::Debug;
 /// Trait implemented by index enums used for uniform buffers.
 pub trait ShaderUniformIndex: Debug + Sized {
     const MAX: usize;
-    fn index(&self) -> u64;
-    fn by_index(index: u64) -> Option<Self>;
+    fn index(&self) -> usize;
+    fn by_index(index: usize) -> Option<Self>;
     fn name() -> &'static str;
-
-    // #[inline]
-    // fn first() -> Self {
-    //     Self::by_index(0).expect("Uniform index type doesn't have any buffers")
-    // }
 }
 
 /// Marker trait for uniform index enums that only contain a single buffer.
